+++ changelog.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Brutal Swarm season
+* Grim operator
+* Grim icon
+
 ## [4.3.0] - 2022-06-28
 
 ### Added